temp dir path passed as argument
